Added a skeleton of the search page for SPEED. Linked to this page from index.js.
ADDED search.js
Search.js is currently a skeleton of the SPEED search page which will let users search for articles based on certain topics and claims.

ADDED search.module.css
search.module.css is the CSS file for the search module. So far it is very basic, it has a bit of padding and it makes the bar at the top blue.

CHANGED index.js
A link to the search page has been added to index.js